refactor: move jobs into a own controller
